Rename ElasticSearch to Elasticsearch
This is how they refer to themselves.
