+++ README.md
@@ -30,10 +30,10 @@ read/write data points from/to InfluxDb.
 
 A collection of probabilistic aggregations such as HyperLogLog.
 
-### [ElasticSearch Connector](elasticsearch) 
+### [Elasticsearch Connector](elasticsearch) 
 
-A Hazelcast Jet connector for ElasticSearch for querying/indexing objects
-from/to ElasticSearch.
+A Hazelcast Jet connector for Elasticsearch for querying/indexing objects
+from/to Elasticsearch.
 
 
 ## Contributing
