+++ elasticsearch/README.md
@@ -1,12 +1,12 @@
-# ElasticSearch Connectors
+# Elasticsearch Connectors
 
-A Hazelcast Jet connector for ElasticSearch (v5..v7) for querying/indexing
-objects from/to ElasticSearch.
+A Hazelcast Jet connector for Elasticsearch (v5..v7) for querying/indexing
+objects from/to Elasticsearch.
 
-There are 3 major versions of ElasticSearch we support:
+There are 3 major versions of Elasticsearch we support:
 
-- [ElasticSearch 5.6.x](elasticsearch-5)
+- [Elasticsearch 5.6.x](elasticsearch-5)
 
-- [ElasticSearch 6.x.x](elasticsearch-6)
+- [Elasticsearch 6.x.x](elasticsearch-6)
 
-- [ElasticSearch 7.x.x](elasticsearch-7)
+- [Elasticsearch 7.x.x](elasticsearch-7)
