+++ elasticsearch/elasticsearch-5/src/test/java/com/hazelcast/jet/elasticsearch/ElasticsearchBaseTest.java
@@ -21,6 +21,9 @@ import com.hazelcast.jet.Jet;
 import com.hazelcast.jet.JetInstance;
 import com.hazelcast.jet.function.FunctionEx;
 import org.apache.http.HttpHost;
+import org.apache.http.auth.UsernamePasswordCredentials;
+import org.apache.http.client.CredentialsProvider;
+import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.elasticsearch.action.get.GetRequest;
 import org.elasticsearch.action.index.IndexRequest;
 import org.elasticsearch.client.RestClient;
@@ -35,24 +38,31 @@ import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Map;
 
+import static org.apache.http.auth.AuthScope.ANY;
 import static org.junit.Assert.assertTrue;
 
-public abstract class ElasticSearchBaseTest {
+public abstract class ElasticsearchBaseTest {
+
+    static final String DEFAULT_USER = "elastic";
+    static final String DEFAULT_PASS = "changeme";
 
     private static final int OBJECT_COUNT = 20;
 
     @Rule
     public ElasticsearchContainer container =
-            new ElasticsearchContainer("docker.elastic.co/elasticsearch/elasticsearch:6.8.0");
+            new ElasticsearchContainer("docker.elastic.co/elasticsearch/elasticsearch:5.6.16");
     JetInstance jet;
     IListJet<User> userList;
     String indexName = "users";
-    private RestHighLevelClient client;
+    private RestClient client;
+    private RestHighLevelClient highLevelClient;
 
     @Before
     public void setupBase() {
         container.start();
-        client = createClient(container.getHttpHostAddress());
+
+        client = createClient(container.getContainerIpAddress(), mappedPort());
+        highLevelClient = new RestHighLevelClient(client);
 
         jet = Jet.newJetInstance();
 
@@ -69,15 +79,25 @@ public abstract class ElasticSearchBaseTest {
         jet.shutdown();
     }
 
+    int mappedPort() {
+        String hostAddress = container.getHttpHostAddress();
+        return Integer.parseInt(hostAddress.split(":")[1]);
+    }
+
     void assertIndexes() throws IOException {
         for (int i = 0; i < OBJECT_COUNT; i++) {
             GetRequest request = new GetRequest(indexName).id(String.valueOf(i));
-            assertTrue(client.exists(request));
+            assertTrue(highLevelClient.exists(request));
         }
     }
 
-    static RestHighLevelClient createClient(String containerAddress) {
-        return new RestHighLevelClient(RestClient.builder(HttpHost.create(containerAddress)));
+    static RestClient createClient(String containerAddress, int port) {
+        CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
+        credentialsProvider.setCredentials(ANY, new UsernamePasswordCredentials(DEFAULT_USER, DEFAULT_PASS));
+
+        return RestClient.builder(new HttpHost(containerAddress, port))
+                         .setHttpClientConfigCallback(httpClientBuilder ->
+                                 httpClientBuilder.setDefaultCredentialsProvider(credentialsProvider)).build();
     }
 
     static FunctionEx<User, IndexRequest> indexFn(String indexName) {
