+++ elasticsearch/elasticsearch-6/src/main/java/com/hazelcast/jet/elasticsearch/package-info.java
@@ -15,6 +15,6 @@
  */
 
 /**
- * Contains sources and sinks for ElasticSearch
+ * Contains sources and sinks for Elasticsearch
  */
 package com.hazelcast.jet.elasticsearch;
