+++ elasticsearch/elasticsearch-6/src/test/java/com/hazelcast/jet/elasticsearch/ElasticsearchBaseTest.java
@@ -21,9 +21,6 @@ import com.hazelcast.jet.Jet;
 import com.hazelcast.jet.JetInstance;
 import com.hazelcast.jet.function.FunctionEx;
 import org.apache.http.HttpHost;
-import org.apache.http.auth.UsernamePasswordCredentials;
-import org.apache.http.client.CredentialsProvider;
-import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.elasticsearch.action.get.GetRequest;
 import org.elasticsearch.action.index.IndexRequest;
 import org.elasticsearch.client.RestClient;
@@ -38,31 +35,24 @@ import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.apache.http.auth.AuthScope.ANY;
 import static org.junit.Assert.assertTrue;
 
-public abstract class ElasticSearchBaseTest {
-
-    static final String DEFAULT_USER = "elastic";
-    static final String DEFAULT_PASS = "changeme";
+public abstract class ElasticsearchBaseTest {
 
     private static final int OBJECT_COUNT = 20;
 
     @Rule
     public ElasticsearchContainer container =
-            new ElasticsearchContainer("docker.elastic.co/elasticsearch/elasticsearch:5.6.16");
+            new ElasticsearchContainer("docker.elastic.co/elasticsearch/elasticsearch:6.8.0");
     JetInstance jet;
     IListJet<User> userList;
     String indexName = "users";
-    private RestClient client;
-    private RestHighLevelClient highLevelClient;
+    private RestHighLevelClient client;
 
     @Before
     public void setupBase() {
         container.start();
-
-        client = createClient(container.getContainerIpAddress(), mappedPort());
-        highLevelClient = new RestHighLevelClient(client);
+        client = createClient(container.getHttpHostAddress());
 
         jet = Jet.newJetInstance();
 
@@ -79,25 +69,15 @@ public abstract class ElasticSearchBaseTest {
         jet.shutdown();
     }
 
-    int mappedPort() {
-        String hostAddress = container.getHttpHostAddress();
-        return Integer.parseInt(hostAddress.split(":")[1]);
-    }
-
     void assertIndexes() throws IOException {
         for (int i = 0; i < OBJECT_COUNT; i++) {
             GetRequest request = new GetRequest(indexName).id(String.valueOf(i));
-            assertTrue(highLevelClient.exists(request));
+            assertTrue(client.exists(request));
         }
     }
 
-    static RestClient createClient(String containerAddress, int port) {
-        CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
-        credentialsProvider.setCredentials(ANY, new UsernamePasswordCredentials(DEFAULT_USER, DEFAULT_PASS));
-
-        return RestClient.builder(new HttpHost(containerAddress, port))
-                         .setHttpClientConfigCallback(httpClientBuilder ->
-                                 httpClientBuilder.setDefaultCredentialsProvider(credentialsProvider)).build();
+    static RestHighLevelClient createClient(String containerAddress) {
+        return new RestHighLevelClient(RestClient.builder(HttpHost.create(containerAddress)));
     }
 
     static FunctionEx<User, IndexRequest> indexFn(String indexName) {
