+++ elasticsearch/elasticsearch-7/src/main/java/com/hazelcast/jet/elasticsearch/ElasticsearchSinks.java
@@ -37,11 +37,11 @@ import java.io.IOException;
 import static org.apache.http.auth.AuthScope.ANY;
 
 /**
- * Contains factory methods for ElasticSearch sinks
+ * Contains factory methods for Elasticsearch sinks
  */
-public final class ElasticSearchSinks {
+public final class ElasticsearchSinks {
 
-    private ElasticSearchSinks() {
+    private ElasticsearchSinks() {
     }
 
     /**
@@ -49,7 +49,7 @@ public final class ElasticSearchSinks {
      * using bulk requests.
      *
      * @param name                Name of the created sink
-     * @param clientSupplier      ElasticSearch rest client supplier
+     * @param clientSupplier      Elasticsearch rest client supplier
      * @param bulkRequestSupplier Bulk request supplier, will be called to obtain a
      *                            new {@link BulkRequest} instance after each call.
      * @param indexFn             Creates an {@link IndexRequest} for each object
