+++ elasticsearch/elasticsearch-7/src/test/java/com/hazelcast/jet/elasticsearch/ElasticsearchBaseTest.java
@@ -42,7 +42,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
-public abstract class ElasticSearchBaseTest {
+public abstract class ElasticsearchBaseTest {
 
     private static final int OBJECT_COUNT = 20;
